convert IX->ID when calculating pseudo IDs (Recidiviz/recidiviz-data#30422)
## Description of the change

This is what was causing impersonation to fail (and would have caused
actual logins to fail) for supervisors in ID during the bug bash today,
because the admin panel roster still uses ID as the state code and so
its pseudo ID calculation was different from the one in the outliers BQ
views, so the backend wasn't able to find the data for the currently
logged in supervisor.

Since pseudo IDs are used in URLs, we want them to be stable, and since
the intention is still to eventually convert US_IX back to US_ID, I
opted to convert IX->ID instead of the other way around.

Made the change for insights, sentencing, and a reference view that
would be used for admin panel roster sync if it were enabled in ID. I
left workflows alone since its URLs are already in use.

Loaded to sandbox at `dana_ids` and confirmed that our two supervisor
TTs now have the same pseudo id in the outliers views and in the product
roster. I didn't check any of the other views.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes https://github.com/Recidiviz/recidiviz-dashboards/issues/5608

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 56616da94f72d7f8d839833ba9c5a12f417f9f2e