[Justice Counts] Double-write agency metrics to the MetricSettings table. (Recidiviz/recidiviz-data#30362)
## Description of the change

Double-write agency metrics to the MetricSettings table.

Zenhub Ticket Recidiviz/recidiviz-data#28420

Whenever agency metric interfaces are written as agency datapoints to
the Datapoint table, we will also write the same metric interfaces to
the MetricSetting table. We will regularly test for parity between
agency metric interfaces stored in the Datapoint table and the
MetricSetting table to confirm that the MetricSetting is a correct
representation of agency metric settings.

We wrap the MetricSetting update logic in a try-except so that if any
bugs or issues occur, they don't cause problems in production. However,
we will also raise an error log in sentry to alert us that the
MetricSetting write-logic has a bug.

## Testing
I deployed a playtesting link and tested that this is double-writing as
intended.
https://brandontest---publisher-web-b47yvyxs3q-uc.a.run.app/

GitOrigin-RevId: 6b1f1fb9597e7766bb777971ad6a5625fecccefc