[Ingest View Tests] Allow comments in fixture CSVs (Recidiviz/recidiviz-data#29100)
## Description of the change

Allow for comments in csvs, starting with 💬.

Comments are not a standard feature of CSVs, [RFC
4180](https://www.loc.gov/preservation/digital/formats/fdd/fdd000323.shtml)
doesn't declare a way to support them, but some CSV parsers allow for
ignoring comments (content on a line after a specified character). We
would not want to use this in raw data import, as it could cause
unintended data loss, but in test data where we control the content, it
is safer to enable commenting and could be useful in adding more clarity
for what each row of data is attempting to test.

This PR proposes using the 💬 character to specify comments as, while it
is difficult to type, it is unlikely to come up in actual data and thus
get in the way of actual testing (as compared to `#`).

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#29115

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 472799ac2f590b83405df3487a8655742a6664d1