[US_ME] Update ResidentRecord for ME to pull from StateIncarcerationSentence (Recidiviz/recidiviz-data#30472)
## Description of the change

We pull release dates from raw data in ME, but the ingested data is more
accurate. We join in the ingested data in resident record, but to keep
the scope small, we only use ingested release dates when the current
release date in resident record is null.

In the end, this only updates 72 rows in resident record. [This query
shows the
changes](https://console.cloud.google.com/bigquery?ws=!1m7!1m6!12m5!1m3!1srecidiviz-staging!2sus-central1!3sfbc5715f-8a91-4653-a5e1-46d07b0f0602!2e1).

Note that this change should only affect what is shown in the front end,
and should not affect TES, as we're modifying resident record directly.
I will make PRs soon to update the upstream views so that TES is
affected, but for launch, we want as small a change as possible.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 3db7ad64ea5aaad15be7dd0c66eee50e11d2fe52