[Workflows] Opportunity Config Admin Panel Work (Recidiviz/recidiviz-data#30364)
## Description of the change

This PR stacks a bunch of work on the workflows config backend/admin
panel. The big change is that you can now edit a complete configuration
in the admin panel using real controls. The layout could use a styling pass
and there's a lot more validation it could be doing, but it works.

Other improvements:
* If the frontend sends a featureVariants parameter and the token
belongs to a recidiviz user, it'll use the frontend's FVs
* You can't confuse the admin panel by changing states while inside an
opportunity
* Snoozes are sent everywhere correctly

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: b3a10ca9a216a1489cc9cc4dccbd8330bd33a857