[Tools] Update initial pipenv setup script (Recidiviz/recidiviz-data#30470)
This PR has a couple small
updates to our initial pipenv setup script that I think will help with
quality of life:

- Checks that the pip being used to install pipenv is from a pyenv shim.
This should prevent folks from trying to run this and install stuff to
their system python.
- Checks that pipenv is installed
- If the architecture is for MacOS, then run the openssl/cmake stuff.
Otherwise continue on.
- I renamed the script now that it checks if running on a mac

Tested this on my Mac and on the debian VM and both worked. I double
checked by running tests on each machine.

GitOrigin-RevId: 68dfdf7fa06ed3476e5847004283596fd3efe567