[Validation] Add mechanism for making attrs validators exempt on a per-state basis (Recidiviz/recidiviz-data#30488)
## Description of the change

Adds a wrapper that can be used with any attrs validator function to
exempt a state or set of states from that validation.

It could be used once @Lesleylambaric's email validators PR is merged to
apply the validation for all states except the ones that still don't
have emails cleaned up.
```
@attr.s(eq=False, kw_only=True)
class StateStaff(
    HasMultipleExternalIdsEntity[StateStaffExternalId],
    RootEntity,
    BuildableAttr,
    DefaultableAttr,
):
   ...
    email: Optional[str] = attr.ib(
        default=None,
        validator=state_exempted_validation(
            attr_validators.opt_is_valid_email, exempted_states={StateCode.US_ME, etc...}
        ),
    )
   ...

```

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Related to Recidiviz/recidiviz-data#29072

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 73759a7c06eec330db2b6984455968354589c3b9