[BigQuery] Check that regex-identified parent tables match sqlglot (Recidiviz/recidiviz-data#30405)
## Description of the change

We use an (imperfect) regex mechanism to determine the parent tables of
each of our BigQuery views which will sometimes miss valid table
references if they aren't surrounded in backticks and will sometimes
identify extra tables that are inside of a comment. I looked into
switching over to use `sqlglot` to parse parent tables but it is ~800x
slower (16s to parse tables for all our views vs .02 seconds for the
regex method). As a result, I don't think we should switch over the main
implementation. Instead, this PR adds a unit test that will fail when
the regex-parsed parent tables don't match what sqlglot finds.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Related to discussion in this thread:
https://recidiviz.slack.com/archives/C028X32LRH7/p1717554284969459

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 7de4affef38d0045e8bada20cbb0ffdceb7f4254