add scheduled query for product roster archive (Recidiviz/recidiviz-data#29972)
## Description of the change

Add a scheduled query to append the `product_roster` view to
`product_roster_archive` (along with an `export_date`) every day. Since
this view isn't already being exported, this seemed easier than the
regular process (exporting it, cloud storage notification, archive
endpoint, view that reads from GCS, view that parses the previous view).

~WIP because `google_bigquery_dataset` wasn't added as a datasource
until version 5.5 of the provider, which means upgrading past 5.0 and I
didn't want to do that unless we know we want to go forward with this
approach.~

@alexabatino, curious how you feel about doing this vs the traditional
way. On one hand, it adds a second way to do something we're used to
doing a different way. On the other hand, it's way fewer moving parts.
If you like it and it ends up working out, we can also talk about
creating scheduled queries like this for our other archive views-
there's a slight difference here in that this will query the data
*regardless* of whether or not it was exported, which we may or may not
be OK with (I haven't thought it through too much yet).

@ohaibbq two questions:
1. Is there a good/easy way to make sure we're alerted if this fails?
There's a parameter that lets you send an email if it fails but the
email goes to the query owner, which in this case would be the service
account.
2. Right now reference_views isn't managed by TF, which is why I went
the datasource route, but should it be? Do you have an opinion?

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#29182
Relevant to Recidiviz/recidiviz-data#30249

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: c055482ada1dabfc1f190174d1758ee0e870cbd6