[US_ND] Add ORDER_TYPE column & reconfigure offenses table (Recidiviz/recidiviz-data#30297)
## Description of the change

We requested that the `ORDER_TYPE` column be added to the
`recidiviz_elite_offences` table transfer. They sent an updated table
with a new (more useful) set of columns, and also a new name
(`elite_offender_offences`). This PR adds the new columns and changes
the name of the table. Since the first transfer of this new table is a
full historical version of the table, I will delete and deprecate the
old version of the table and its metadata. This file is not used
anywhere downstream, so no references will need to be updated.

I successfully imported this raw data to a sandbox at
`recidiviz-staging.elise_us_nd_raw_data.elite_offender_offences`

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 355c608d47e94541a23d2ce615c51aa2321641f1