[Security] Pin GH Actions to particular checksum (Recidiviz/recidiviz-data#30291)
## Description of the change

> This PR pins our GH actions to full-length commit SHA per Github's
[recommendations](https://docs.github.com/en/actions/security-guides/security-hardening-for-github-actions#using-third-party-actions).
Per Github's documentation, `Pinning to a particular SHA helps mitigate
the risk of a bad actor adding a backdoor to the action's repository, as
they would need to generate a SHA-1 collision for a valid Git object
payload.` This recent [security
audit](https://pin-gh-actions.kammel.dev/?utm_source=tldrsec.com&utm_medium=newsletter&utm_campaign=tl-dr-sec-233-awesome-detection-engineering-security-gpts-how-to-build-a-cybersecurity-start-up)
of the top GH repos prompted this change. It revealed that only 2% of
the most popular 10k repos are doing this.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

https://github.com/Recidiviz/zenhub-tasks/issues/1685

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 77ac6531c4f65f8bf8323c5b30c7bca41d0c90a6